change memory variables datatype

+++ Config.cpp
@@ -32,10 +32,10 @@ Config loadConfig(const std::string& filePath) {  // ✅ No default here
         else if (key == "delays-per-exec") config.delaysPerExec = std::stoull(value);
 
         // MO2 NEW FIELDS
-        else if (key == "max-overall-mem") config.maxOverallMemory = std::stoull(value);
-        else if (key == "mem-per-frame") config.memPerFrame = std::stoull(value);
-        else if (key == "min-mem-per-proc") config.minMemPerProcess = std::stoull(value);
-        else if (key == "max-mem-per-proc") config.maxMemPerProcess = std::stoull(value);
+        else if (key == "max-overall-mem") config.maxOverallMemory = static_cast<size_t>(std::stoull(value));
+        else if (key == "mem-per-frame") config.memPerFrame = static_cast<size_t>(std::stoull(value));
+        else if (key == "min-mem-per-proc") config.minMemPerProcess = static_cast<size_t>(std::stoull(value));
+        else if (key == "max-mem-per-proc") config.maxMemPerProcess = static_cast<size_t>(std::stoull(value));
     }
 
     return config;
